move MemoryStream/CounterStream to MeshUtils
