+++ Plugin~/MeshUtils.vcxproj.filters
@@ -79,6 +79,9 @@
     <ClInclude Include="MeshUtils\muAlgorithm.h">
       <Filter>MeshUtils</Filter>
     </ClInclude>
+    <ClInclude Include="MeshUtils\muStream.h">
+      <Filter>MeshUtils</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="MeshUtils">
@@ -117,6 +120,9 @@
       <Filter>CrashReporter</Filter>
     </ClCompile>
     <ClCompile Include="MeshUtils\muCompression.cpp">
+      <Filter>MeshUtils</Filter>
+    </ClCompile>
+    <ClCompile Include="MeshUtils\muStream.cpp">
       <Filter>MeshUtils</Filter>
     </ClCompile>
   </ItemGroup>
